+++ changelog.md
@@ -1,3 +1,10 @@
 # ZealousDateAndTime Changelog
+## v1.0.1
+- Added languages for month names <cl>(Jan/Feb/Mar/etc.)</c> and day of week names <cl>(Sun/Mon/Tue/etc.)</c>.
+  - Only romanizations are available for Japanese, Korean, and Chinese.
+  - Accent marks from most languages (including above-mentioned romanizations) are not present. (Fonts in Geometry Dash only use the 26 letters of the English alphabet.)
+  - If any translations are incorrect, let me know by pressing the Discord button in the lower left corner.
+  - Languages besides English include: <cl>Spanish</c>, <cl>French</c>, <cl>Italian</c>, <cl>German</c>, <cl>Portuguese</c>, <cl>Finnish</c>, <cl>Czech</c>, <cl>Slovak</c>, <cl>Croatian</c>, <cl>Romanian</c>, <cl>Dutch</c>, <cl>Welsh</c>, <cl>Scots Gaelic</c>, <cl>Indonesian</c>, <cl>Japanese (Romaji)</c>, <cl>Chinese Mandarin (Pinyin)</c>, and <cl>Korean (Revised Romanization)<cl>.
+- Allow adjusting how much a month's name of the day of the week's name is truncated.
 ## v1.0.0
 - Initial release (on GitHub).